perf: 执行失败场景下的临时文件清理策略优化 #592
超出指定容量则清理最旧的文件

+++ src/backend/job-file-worker-sdk/service-job-file-worker-sdk/src/main/java/com/tencent/bk/job/file/worker/config/WorkerConfig.java
@@ -70,7 +70,7 @@ public class WorkerConfig {
     private String downloadFileDir;
 
     @Value("${job.file-worker.download-file.max-size-gb:100}")
-    private Integer maxSizeGB;
+    private Long maxSizeGB;
 
     @Value("${job.file-worker.download-file.expire-days:7}")
     private Integer downloadFileExpireDays;
